requirements.txt
# Core dependencies
streamlit==1.28.0
python-dotenv==1.0.0

# AI integration (Mistral AI)
mistralai==1.5.1  # Latest available version

# Data processing
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1

# NLP and ML
tensorflow==2.14.0
transformers==4.34.1
nltk==3.8.1
spacy==3.7.1
gensim==4.3.2

# Visualization
matplotlib==3.8.0
plotly==5.17.0
wordcloud==1.9.2

# PDF processing
pdfminer.six==20221105
python-docx==1.0.1

# Web scraping (for job data)
beautifulsoup4==4.12.2
requests==2.31.0

# Testing
pytest==7.4.2
